+++ Wms12m.Presentation/Reports/SatınalmaSiparisFormu.cs
@@ -11,21 +11,20 @@ namespace Wms12m
 {
     public static class SatınalmaSiparisFormu
     {
-        public static void SatinalmaSiparisFormu(string sipEvrakNo, string hesapKodu, int siparisTarih, bool pdfKaydet, string SirketKodu)
+        public static void SatinalmaSiparisFormu(string sipEvrakNo, string hesapKodu, int siparisTarih, bool pdfKaydet, string sirketKodu)
         {
-            WMSEntities db = new WMSEntities();
-            if (string.IsNullOrEmpty(sipEvrakNo) || string.IsNullOrEmpty(hesapKodu))
-                return;
+            var db = new WMSEntities();
+            if (string.IsNullOrEmpty(sipEvrakNo) || string.IsNullOrEmpty(hesapKodu)) return;
 
-            DsSatinalma dset = new DsSatinalma();
-            string dvzCinsi = "TL";
+            var dset = new DsSatinalma();
+            var dvzCinsi = "TL";
             short islemTip = 1;
-            KKP_CHK chk = new KKP_CHK();
-            SatTalep talep = new SatTalep();
-            bool hata = false;
-            int say = 1;
+            var chk = new KKP_CHK();
+            var talep = new SatTalep();
+            var hata = false;
+            var say = 1;
             DateTime? sipTarih = null;
-            SatinalmaRaporData _FDD = new SatinalmaRaporData()
+            var fDD = new SatinalmaRaporData()
             {
                 FTDD = new List<RaporFTD>(),
                 CHKK = new List<KKP_CHK>(),
@@ -33,30 +32,28 @@ namespace Wms12m
                 TLP = new List<SatTalep>()
             };
             decimal ftdMalBedeli = 0, ftdKDV = 0, ftdToplam = 0;
-            var RPR = db.MultipleResults(string.Format("[FINSAT6{0}].[wms].[Satinalma_RaporDetay] @EvrakNo='{1}',@HesapKodu='{2}',@Siptarih={3}", SirketKodu, sipEvrakNo, hesapKodu, siparisTarih)).With<RaporFTD>().With<RaporSiparis>().With<KKP_CHK>().With<SatTalep>().Execute();
-            foreach (RaporFTD item in RPR[0])
+            var rPR = db.MultipleResults(string.Format("[FINSAT6{0}].[wms].[Satinalma_RaporDetay] @EvrakNo='{1}',@HesapKodu='{2}',@Siptarih={3}", sirketKodu, sipEvrakNo, hesapKodu, siparisTarih)).With<RaporFTD>().With<RaporSiparis>().With<KKP_CHK>().With<SatTalep>().Execute();
+            foreach (RaporFTD item in rPR[0])
             {
-                _FDD.FTDD.Add(item);
+                fDD.FTDD.Add(item);
                 if (item.DovizCinsi.ToString().Trim() != "")
                     dvzCinsi = item.DovizCinsi.ToString().Trim();
 
-                short satirTip = (short)item.SatirTip;
-                decimal tutar = (decimal)item.Tutar;
+                var satirTip = (short)item.SatirTip;
+                var tutar = (decimal)item.Tutar;
 
-                if (satirTip == 0)
-                    ftdMalBedeli = tutar;
-                else if (satirTip == 9)
-                    ftdKDV = tutar;
-                else if (satirTip == 12)
-                    ftdToplam = tutar;
+                if (satirTip == 0) ftdMalBedeli = tutar;
+                else if (satirTip == 9) ftdKDV = tutar;
+                else if (satirTip == 12) ftdToplam = tutar;
             }
-            foreach (RaporSiparis item in RPR[1])
+
+            foreach (RaporSiparis item in rPR[1])
             {
-                _FDD.SPII.Add(item);
+                fDD.SPII.Add(item);
                 sipTarih = new DateTime(1900, 1, 1).AddDays((int)item.Tarih - 2);
                 islemTip = (short)item.IslemTip;
 
-                DsSatinalma.SiparisRow row = dset.Siparis.NewSiparisRow();
+                var row = dset.Siparis.NewSiparisRow();
                 row.SatirNo = say;
                 row.Malzeme = item.MalKodu.ToString();
                 row.DvzCinsi = item.DovizCinsi.ToString() == "" ? "TL" : item.DovizCinsi.ToString();
@@ -85,22 +82,23 @@ namespace Wms12m
 
                 row.DovizCinsi = dvzCinsi;
 
-
-
                 dset.Siparis.AddSiparisRow(row);
 
                 say++;
             }
-            foreach (KKP_CHK item in RPR[2])
+
+            foreach (KKP_CHK item in rPR[2])
             {
-                _FDD.CHKK.Add(item);
+                fDD.CHKK.Add(item);
                 chk = (item);
             }
-            foreach (SatTalep item in RPR[3])
+
+            foreach (SatTalep item in rPR[3])
             {
-                _FDD.TLP.Add(item);
+                fDD.TLP.Add(item);
                 talep = item;
             }
+
             try
             {
                 foreach (var item in dset.Siparis)
@@ -108,28 +106,25 @@ namespace Wms12m
                     if (talep.IstenenTarih != null)
                         item.IstenenTarih = ((DateTime)talep.IstenenTarih).ToString("dd.MM.yyyy");
                 }
-
             }
             catch (Exception)
             {
                 hata = true;
-
             }
-            if (hata)
-                return;
+
+            if (hata) return;
 
             try
             {
-
-                GenelAyarVeParam satisUzmani = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=0 AND SiparisSorumlu = '{0}'", talep.Kaydeden)).FirstOrDefault();
-                GenelAyarVeParam gmy = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=1 AND SiparisSorumlu = '{0}'", talep.GMYMaliOnaylayan)).FirstOrDefault();
+                var satisUzmani = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=0 AND SiparisSorumlu = '{0}'", talep.Kaydeden)).FirstOrDefault();
+                var gmy = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=1 AND SiparisSorumlu = '{0}'", talep.GMYMaliOnaylayan)).FirstOrDefault();
 
                 if (islemTip == (short)KKPIslemTipSPI.İçPiyasa)
                 {
-                    SatSipForm rep = new SatSipForm();
+                    var rep = new SatSipForm();
                     if (satisUzmani != null)
                     {
-                        User user = db.Users.Where(m => m.Kod == satisUzmani.SiparisSorumlu).FirstOrDefault();
+                        var user = db.Users.Where(m => m.Kod == satisUzmani.SiparisSorumlu).FirstOrDefault();
                         if (user != null)
                         {
                             rep.lblBirinciKisiAd.Text = user.AdSoyad;
@@ -139,7 +134,7 @@ namespace Wms12m
 
                     if (gmy != null)
                     {
-                        User user = db.Users.Where(m => m.Kod == gmy.SiparisSorumlu).FirstOrDefault();
+                        var user = db.Users.Where(m => m.Kod == gmy.SiparisSorumlu).FirstOrDefault();
                         if (user != null)
                         {
                             rep.PicikinciKisiImza.Image = ByteArrayToImage.byteArrayToImage(gmy.SiparisSorumluImza);
@@ -148,10 +143,10 @@ namespace Wms12m
 
                     if (talep.GMOnaylayan != null)
                     {
-                        GenelAyarVeParam gm = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=2 AND SiparisSorumlu = '{0}'", (talep.GMOnaylayan ?? "-1"))).FirstOrDefault();
+                        var gm = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=2 AND SiparisSorumlu = '{0}'", (talep.GMOnaylayan ?? "-1"))).FirstOrDefault();
                         if (gm != null)
                         {
-                            User user = db.Users.Where(m => m.Kod == gm.SiparisSorumlu).FirstOrDefault();
+                            var user = db.Users.Where(m => m.Kod == gm.SiparisSorumlu).FirstOrDefault();
                             if (user != null)
                             {
                                 rep.PicUcuncuKisiImza.Image = ByteArrayToImage.byteArrayToImage(gm.SiparisSorumluImza);
@@ -161,10 +156,8 @@ namespace Wms12m
                     else
                         rep.lblUcuncuKisiUnvn.Visible = false;
 
-
-
                     rep.lblUnvan.Text = chk.Unvan1;
-                    rep.lblAdrs.Text = String.Format("{0} {1} {2}", chk.FaturaAdres1, chk.FaturaAdres2, chk.FaturaAdres3);
+                    rep.lblAdrs.Text = string.Format("{0} {1} {2}", chk.FaturaAdres1, chk.FaturaAdres2, chk.FaturaAdres3);
                     rep.lblTel.Text = chk.Telefon1;
                     rep.lblFax.Text = chk.Fax1;
                     rep.lblEmail.Text = chk.SatAlmaIslemEMail;
@@ -176,21 +169,18 @@ namespace Wms12m
 
                     if (pdfKaydet)
                     {
-                        string temp = String.Format("{0}{1}.pdf", Path.GetTempPath(), sipEvrakNo);
-                        if (File.Exists(temp))
-                        {
-                            File.Delete(temp);
-                        }
+                        var temp = string.Format("{0}{1}.pdf", Path.GetTempPath(), sipEvrakNo);
+                        if (File.Exists(temp)) File.Delete(temp);
                         rep.ExportToPdf(temp);
                     }
                 }
                 else if (islemTip == (short)KKPIslemTipSPI.DışPiyasa)
                 {
-                    SatSipForm rep = new SatSipForm();
+                    var rep = new SatSipForm();
 
                     if (satisUzmani != null)
                     {
-                        User user = db.Users.Where(m => m.Kod == satisUzmani.SiparisSorumlu).FirstOrDefault();
+                        var user = db.Users.Where(m => m.Kod == satisUzmani.SiparisSorumlu).FirstOrDefault();
                         if (user != null)
                         {
                             rep.lblBirinciKisiAd.Text = user.AdSoyad;
@@ -200,7 +190,7 @@ namespace Wms12m
 
                     if (gmy != null)
                     {
-                        User user = db.Users.Where(m => m.Kod == gmy.SiparisSorumlu).FirstOrDefault();
+                        var user = db.Users.Where(m => m.Kod == gmy.SiparisSorumlu).FirstOrDefault();
                         if (user != null)
                         {
                             rep.PicikinciKisiImza.Image = ByteArrayToImage.byteArrayToImage(gmy.SiparisSorumluImza);
@@ -209,10 +199,10 @@ namespace Wms12m
 
                     if (talep.GMOnaylayan != null)
                     {
-                        GenelAyarVeParam gm = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=2 AND SiparisSorumlu = '{1}'", SirketKodu, (talep.GMOnaylayan ?? "-1"))).FirstOrDefault();
+                        var gm = db.Database.SqlQuery<GenelAyarVeParam>(string.Format("SELECT * FROM [Kaynak].[sta].[GenelAyarVeParams] WHERE Tip=1 AND Tip2=2 AND SiparisSorumlu = '{1}'", sirketKodu, (talep.GMOnaylayan ?? "-1"))).FirstOrDefault();
                         if (gm != null)
                         {
-                            User user = db.Users.Where(m => m.Kod == gm.SiparisSorumlu).FirstOrDefault();
+                            var user = db.Users.Where(m => m.Kod == gm.SiparisSorumlu).FirstOrDefault();
                             if (user != null)
                             {
                                 rep.PicUcuncuKisiImza.Image = ByteArrayToImage.byteArrayToImage(gm.SiparisSorumluImza);
@@ -222,10 +212,8 @@ namespace Wms12m
                     else
                         rep.lblUcuncuKisiUnvn.Visible = false;
 
-
-
                     rep.lblUnvan.Text = chk.Unvan1;
-                    rep.lblAdrs.Text = String.Format("{0} {1} {2}", chk.FaturaAdres1, chk.FaturaAdres2, chk.FaturaAdres3);
+                    rep.lblAdrs.Text = string.Format("{0} {1} {2}", chk.FaturaAdres1, chk.FaturaAdres2, chk.FaturaAdres3);
                     rep.lblTel.Text = chk.Telefon1;
                     rep.lblFax.Text = chk.Fax1;
                     rep.lblEmail.Text = chk.SatAlmaIslemEMail;
@@ -237,13 +225,9 @@ namespace Wms12m
 
                     if (pdfKaydet)
                     {
-                        string temp = String.Format("{0}{1}.pdf", Path.GetTempPath(), sipEvrakNo);
-                        if (File.Exists(temp))
-                        {
-                            File.Delete(temp);
-                        }
+                        var temp = string.Format("{0}{1}.pdf", Path.GetTempPath(), sipEvrakNo);
+                        if (File.Exists(temp)) File.Delete(temp);
                         rep.ExportToPdf(temp);
-
                     }
                 }
             }
